fixed and added optional login for /items

+++ resources/item.py
@@ -1,7 +1,7 @@
 import sqlite3
 from flask_restful import Resource, reqparse
 # from flask_jwt import jwt_required
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 
@@ -54,6 +54,11 @@ class Item(Resource):
 
         connection.commit()
         connection.close()"""
+        claims = get_jwt()
+        print(claims)
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required!'}
+            
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -78,25 +83,17 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-    @jwt_required()
+    @jwt_required(optional=True)
     def get(self):
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        print(user_id)
+        if user_id:
+            return {'items': items}, 200
         # make queries only from the Model, not from Resource, makes it too heavy!
-        return {'item': [item.json() for item in ItemModel.find_all()]}
+        return {
+            'items': [item['name'] for item in items],
+            'message': "More data available if you log in."
+            }, 200
         # return {'item': list(map(lambda x: x.json(), ItemModel.query.all()))}
-
-    """@classmethod
-    def get_all_items(cls):
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        query = "SELECT * FROM items"
-        result = cursor.execute(query)
-        rows = result.fetchall()
-        print(rows)
-        connection.close()
-
-        if len(rows) != 0:
-            items = [{'item': item[0], 'price': item[1]} for item in rows]
-            return {'item': items}"""
-
